% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogcc.R
\name{getLabels}
\alias{getLabels}
\title{Get the output labels (getLabels)}
\usage{
getLabels(model = "types", d = "RSEM")
}
\arguments{
\item{model}{is a prediction model name from \code{\link{ogcc}}. See \code{model} argument.}
}
\value{
a character vector of the class labels a model was trained on and can predict.
}
\description{
\code{getLabels} print the levels of the output variable a model has had trained on.
}
\examples{
labels <- getLabels("normal_tumor")
print(labels)

}
\author{
Ayman Abuelela; ayman.elkhodiery@kaust.edu.sa
}
\seealso{
\code{\link{ogcc}} and \code{\link{getMSF}}
}

