% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogcc.R
\name{ogcc}
\alias{ogcc}
\title{O-glycan Cancer Classifier (OGCC)}
\usage{
ogcc(x, model = "types", output = "raw")
}
\arguments{
\item{x}{a numeric data frame or matrix of expression data where columns are the features and rows are cases/samples. The set of features must contain the minimum set of features for the specified model. See `Details`.}

\item{model}{a character string indicating the model to perform the prediction task. It must be one of the following: "normal_tumor", "normal_tumor_reduced", "normal_tumor_types", "normal_tumor_types_reduced", "types" and "types_reduced".}

\item{output}{a character value indicating the form of prediction output of the model. It must be either "raw" or "probs".}
}
\value{
a character in case of output = "raw" or a dataframe of class probalibilities in case of output = "probs"
}
\description{
\code{ogcc} performs a group of cancer class predictions based on the expression of O-glycan-forming glycosyltransferase genes.
}
\details{
\code{ogcc} classifies cancer samples according to one of a set of predefined classification models using O-glycan-forming genes (OGFGs).

The minimum set of OGFGs required in any model can be displayed using \code{\link{getMSF}} function. The required feature for a specific model may differ according to the feature selection step of that model.

To display the output labels the model was trained on and can predict, use the function \code{\link{getLabels}}. These models are not generic models. They were developed and trained on specific set of cancer classes. Going beyond the classes specified for each model may result in misclassification problems.

\code{ogcc} can perform the following classification tasks based on OGFGTs expression profile:

\enumerate{
      \item model \strong{\code{normal_tumor}}: predicts whether a sample is normal or cancer based on the expression profile of the OGFGs.
      \item model \strong{\code{normal_tumor_reduced}}: the same as model \code{normal_tumor} but the features were reduced to the minimum set to bring 95\% accuracy.
      \item model \strong{\code{normal_tumor_types}}: predicts whether the sample is normal or cancer in addition tissue type prediction.
      \item model \strong{\code{normal_tumor_types_reduced}}: the same as model \code{normal_tumor_types} but the features were reduced to the minimum set to bring 95\% accuracy.
      \item model \strong{\code{types}}: predicts the cancer type.
      \item model \strong{\code{types_reduced}}: the same as model \code{types} but the features were reduced to the minimum set to bring 95\% accuracy.
}
}
\author{
Ayman Abuelela; ayman.elkhodiery@kaust.edu.sa
}
\seealso{
\code{\link{getMSF}} and \code{\link{getLabels}}
}

